+++ app/src/main/java/com/cajetan/youtubeplus/data/VideoDatabase.kt
@@ -5,7 +5,7 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import android.content.Context
 
-@Database(entities = [VideoData::class, PlaylistData::class], version = 1, exportSchema = false)
+@Database(entities = [VideoData::class, PlaylistData::class], version = 2, exportSchema = false)
 abstract class VideoDatabase : RoomDatabase() {
     abstract fun videoDataDao(): VideoDataDao
     abstract fun playlistDataDao(): PlaylistDataDao
@@ -18,7 +18,8 @@ abstract class VideoDatabase : RoomDatabase() {
                 synchronized(VideoDatabase::class) {
                     if (INSTANCE == null) {
                         INSTANCE = Room.databaseBuilder(context.applicationContext,
-                                VideoDatabase::class.java, "video_database").build()
+                                VideoDatabase::class.java, "video_database")
+                                .fallbackToDestructiveMigration().build()
                     }
                 }
             }